backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
python-dotenv
pandas
scikit-learn
statsmodels
gunicorn
